// CryptoCurrency-Spam-Detection--frontend/backend/src/shared/enums/chain.enum.ts
export enum Chain {
  ETHEREUM = 'ethereum',
  BITCOIN = 'bitcoin',
  POLYGON = 'polygon',
  ARBITRUM = 'arbitrum',
  OPTIMISM = 'optimism',
  BSC = 'bsc',
  AVALANCHE = 'avalanche',
  SOLANA = 'solana',
  SEPOLIA = 'sepolia',
  FANTOM = 'fantom',
  BASE = 'base',
  CELO = 'celo',
  GNOSIS = 'gnosis',
  CRONOS = 'cronos',
  MOONBEAM = 'moonbeam',
  METIS = 'metis',
  KAVA = 'kava',
}